Revert "feat(entities-plugins): rewrite service protection form [KM-1265]"
This reverts commit 6d83225c558dd43c834510848048531f9cb5a92b.
