+++ packages/entities/entities-plugins/src/components/free-form/shared/RedisSelector.vue
@@ -311,7 +311,6 @@ onBeforeMount(() => {
   if (props.defaultRedisConfigItem || partialValue?.value) {
     const defaultRedisConfigItem = props.defaultRedisConfigItem || partialValue?.value?.[0].id
     redisConfigSelected(defaultRedisConfigItem)
-    usePartial.value = true
   } else if (redisFieldsValue?.value) {
     redisFieldsSaved.value = redisFieldsValue.value
   }
