+++ packages/entities/entities-plugins/src/types/plugins/form-schema.ts
@@ -31,8 +31,6 @@ export interface FieldSchema {
 export interface StringFieldSchema extends FieldSchema {
   type: 'string'
 
-  auto?: boolean
-
   encrypted?: boolean
 
   one_of?: string[]
